+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- migrated dependencies
+- updated dependencies
+- :boom: changed to own reactstrap fork with react 19 support
 
 ## [5.15.0] - 2026-01-20